# server/model_response_model.go
/*
 * Pseudo Service
 *
 * Deterministic or pseudo-random data generator
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

import "github.com/bgadrian/pseudoservice/users"

type ResponseModel struct {

	// Number that was used to generate these users
	Seed int64 `json:"seed"`

	// Use this to get the next users from the deterministic series
	Nextseed int64 `json:"nextseed"`

	Users []*users.User `json:"users"`
}