# users/model_user.go
/*
 * Pseudo Service
 *
 * Deterministic or pseudo-random data generator
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package users

type User struct {

	// UUID for this user
	Id string `json:"id"`

	// First and Last name
	Name string `json:"name"`

	// Gender one of [M, F, ?]
	Gender string `json:"gender,omitempty"`

	Email string `json:"email,omitempty"`

	Country string `json:"country,omitempty"`

	Timezone string `json:"timezone,omitempty"`

	// Random list of UUIDs from the same response. To increase the chance of having friends request for bigger batches.
	Friends []string `json:"friends,omitempty"`
}

func GenerateUsers(seed int64, count int) ([]*User, int64, error) {
	return nil, 0, nil
}